% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_data.R
\name{resample_data}
\alias{resample_data}
\title{Resample Input Data}
\usage{
resample_data(data2, dir1)
}
\arguments{
\item{data2}{Input data by ID. The input dataset to be resampled or transformed.}

\item{dir1}{Temporal directory. Path to the folder where the resampled data will be saved.}
}
\value{
A resampled version of "data2" named "_exp.txt" saved in the directory specified by "dir1."
}
\description{
"resample_data" processes the data based on predefined rules before being read by the C model, and saves the output in a specified directory.
}
\details{
This function takes two arguments: a dataset ("data2") and the temporal directory path ("dir1").
}
